Adds support for Windows ArgsEscaped images
Adds support for Windows container images built by Docker
that contain the ArgsEscaped boolean in the ImageConfig. This
is a non-OCI entry that tells the runtime that the Entrypoint
and/or Cmd are a single element array with the args pre-escaped
into a single CommandLine that should be passed directly to
Windows rather than passed as an args array which will be
additionally escaped.

Signed-off-by: Justin Terry <jlterry@amazon.com>
(cherry picked from commit de3d9993f5d2e4a2fe124b7572ea565c5caeb674)
Signed-off-by: Kirtana Ashok <Kirtana.Ashok@microsoft.com>