runtime: revert part of #6206
In container bundle, the `runtime` file has been reserved for runtime's
`BinaryName` like runc or crun[1]. In #6206, we supported to use
absolute path as shim runtime name and also stored it in `runtime`
file[2]. It will be conflict if user wants to use BinaryName instead of
empty default value.

Before we figure out new filename for shim binary path, I would like to
revert the code from #6206 because the information has been stored in
container store service. If the shim runtime is absolute binary path,
we can just read it from container store service. If it is normal shim name
like `io.containerd.runc.v2`, we need to resolve it into binary path again,
which is the same to before pr #6206.

Reference:

[1]: https://github.com/containerd/containerd/commit/6bcbf88f82e814f76ede351f48b57613540af425
[2]: https://github.com/containerd/containerd/pull/6206

Fixes: #6534

Signed-off-by: Wei Fu <fuweid89@gmail.com>