docs/getting-started.md: massive update
The previous documentation was too much forcusing on the Go API and not useful
for users who are not interested in implementing their own containerd client.
It was also recommending the deprecated way (cri-containerd-*.tar.gz) to install
containerd and its dependencies.

The new documentation recommends the current official way to install containerd,
and provides several links for end users.

This will replace the content of https://containerd.io/docs/getting-started/
after merging the containerd/containerd.io PR 120.

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>