[release/1.6] go.mod: update image-spec to merge-commit of v1 into main
This is a follow-up to 7ede40c5ca3c8d0565798fa653402357b8c088a0, where we pinned
the version of this dependency to prevent go modules rolling it back.

This patch updates the version to use the merge-commit that merged the v1.0
release branch back to the main branch in github.com/opencontainers/image-spec,
so that go modules considers it "more recent" and doesn't roll back.

Updating does not introduce changes in the vendored code, as changs of the merged
commits were either already in the main branch, or only affected non-code files.

- full diff: https://github.com/opencontainers/image-spec/compare/693428a734f5...c5a74bcca799
- raw diff: https://github.com/opencontainers/image-spec/compare/693428a734f5..c5a74bcca799

Thanks to Tõnis Tiigi for pointing to this commit!

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>