shim: fix debug flag not working
As we know, shim starts twice. The first time we execute the shim with the
`start` arguments, the `-debug` argument is passed to the shim
process correctly. But the second time we execute shim, the debug flag
is ignored.

Signed-off-by: wllenyj <wllenyj@linux.alibaba.com>
(cherry picked from commit f710505f4a2760424258a4d517bc261d5bbfce1c)
Signed-off-by: Brian Goff <cpuguy83@gmail.com>