Allow configuring CRI logger read buffer size
CRI logger uses a bufio Reader while redirecting container logs to a log
file. The buffer size for the Reader is currently set to 4KB as default and
is not configurable. After this change, the buffer size can be configured
through the CRI Plugin config and will enable clients to allocate larger
buffers if needed to improve IO performance. It also sets 4KB as the
default minimum size for the buffer. So, if the configured value is less
than 4KB, we continue to allocate the default minimum.

Signed-off-by: Swagat Bora <swagatbora90@gmail.com>