Depend on pre-installed go-winres for go:generate calls
This fixes cross-compliation, because `go run` builds binaries for the
_target_ when cross-compiling, which is pretty useless when used in `go
generate`.

Unless this[1] has changed in the last few years, the best response I've
seen about this is basically "Punt it over to your user to manage the
necessary tools". It'd be nice if `go run` had some kind of escape-hatch
to "Build for the host" to cover this use-case, as I feel like the
existing escape-hatch of "provide an emulator host" is too much work for
this use-case.

[1] https://github.com/golang/go/issues/33354#issuecomment-516525307

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>