Add validations for Windows HostProcess CRI configs
HostProcess containers require every container in the pod to be a
host process container and have the corresponding field set. The Kubelet
usually enforces this so we'd error before even getting here but we recently
found a bug in this logic so better to be safe than sorry.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>
(cherry picked from commit 978ff393d2544f729284cd2dfb70e497c4038a7c)
Signed-off-by: Daniel Canter <dcanter@microsoft.com>