DockerResolver: support immutable tags
The tag part in `example.com/hello-world:42.0@sha256:deadbeef` had been just ignored.
The tag was never verified with the specified digest because the tag was considered to be mutable.
Even the existence of the tag was never verified, and it was causing a lot of confusions like issue 6450.

This commit enables verification of the tag string that matches the `ImmutableTagMatcher` function.
The matcher functions defaults to `HasSuffix(tag, "-immutable") || HasSuffix(tag, "_immutable") || HasSuffix(tag, ".immutable")`.

e.g., `example.com/hello-world:42-immutable@sha256:deadbeef` is a ref string with an immutable tag.

When the tag matches, the resolver always verifies the existence of the tag, and also verifies the digest
if the digest is present in the ref string.

i.e., pulling `example.com/hello-world:42-immutable@sha256:deadbeef` now fails if "42-immutable" does not exist
or exists with an unexpected digest on `example.com/hello-world`.

The matcher convention is being submitted to the OCI Distribution Spec in opencontainers/distribution-spec PR 320.

Fixes issue 6450

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>