converter: improve concurrency performance
Introduce `converter.WithSingleflight` for `converter.New` and
`converter.NewDefault`, it uses singleflight to make sure that
only one conversion execution is in-flight for a given child
descriptor at a time.

This change can improve converter performance to avoid getting lock
on writing to the content store, in particular, in the use case of implementing
an image conversion service based on the converter (e.g. [accled](https://github.com/goharbor/acceleration-service)).

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>