release workflow: increase timeout to 30 minutes
In the 1.6.7 release, we saw significantly longer execution time for
producing builds that exceeded the previous timeout of 10 minutes,
causing the workflow to fail.  After increasing to 20 minutes in the
release/1.6 branch, we continued to see one failure (which succeeded on
retry).

Increase to 30 minutes to provide additional buffer for the build to
complete.

Signed-off-by: Samuel Karp <samuelkarp@google.com>
(cherry picked from commit f8add9263a9ba64ad4400b7c1877dbd527963fbb)
Signed-off-by: Samuel Karp <samuelkarp@google.com>