integration: TestUpdateContainerResources_MemoryLimit: remove TODO comment
This comment was added in 09a0c9471bf687ebcc7d8f8594606d127b3a803e when the
Windows integration tests were enabled. The PR (microsoft/hcsshim#931) was
merged, and part of hcsshim v0.9.0, and support for resource limits on Windows
was added in 2bc77b8a2865efe83bc50225223555a919191f54, so it looks like this
comment is no longer current.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>