image/label: print more characters of label keys
Like stargz and nydus remote snapshotter, some snapshots
lables are introduced and passed to snapshotter from containerd
automatically. The label keys' length are all longer than 10.

The limitation of 10 characters makes it harder to debug what label
is not appropriate. So we'd better to print more of the wrong label.

Signed-off-by: Changwei Ge <gechangwei@bytedance.com>