support fetching containerd from non public GCS buckets
- add support to fetch and download containerd tarball from GCS buckets
that require authentication.

GCS_BUCKET_TOKEN should have read access to the bucket from which
artifacts are to be fetched. The token is expected to be present in
the instance metadata of the VM, similar to other node environment
variables

Signed-off-by: Akhil Mohan <makhil@vmware.com>
(cherry picked from commit 9df96dc46acf2108b2e299553cceaa3dac1ee2ad)
Signed-off-by: Akhil Mohan <akhilerm@gmail.com>