daemon: add drop-in config fragments for custom daemon config
Motivation:

Some components, such as snapshotter needs to configure the containerd
config file dynamically, it's not so easy to modify the config file,
especially by Linux command such as `sed` because the toml file has
its hierarchy relations, also it has comments in the file, so simply
replacing or appending is not so easy to be used to update
the containerd's config file.

Components write their config fragments, and letting them be imported
by the `import` instrument is an ideal resolution, but modifying
the `imports` instrument has the same problem as above, it's hard
to edit it, for example, add, delete or change the order of the files
in the import list.

Solution:

This commit will add the drop-in config fragments like other traditional
Linux packages to containerd. Users can add their configs into the
`config.d` directory along with the main config file(normally `config.toml`)
but need not modify the main config file.

For example, if we have a config file like this:

```toml
imports = ["import-1.toml", "import-2.toml"]
```

And drop-in configs in `config.d`:

```
10-plugin-1.toml
20-plugin-2.toml
12-plugin-3.toml
```

The drop-ins will be processed seems to be added to the `imports` item in the
main config file. The result equals:

```toml
imports = ["import-1.toml", "import-2.toml", "10-plugin-1.toml", "12-plugin-3.toml", "20-plugin-2.toml"]
```

And the drop-ins will have high priority than imported files.

Signed-off-by: bin liu <liubin0329@gmail.com>