Platform matcher: allow for running of newer Win images under Hyper-V.
Following the addition of Hyper-V isolation, it should be possible for
Windows hosts to run container images which are older or newer than
the host build if the Windows host is version RS5 (1809) or above.

This patch updates the Windows platform matching logic to allow for
build number mismatches between Windows images and hosts iff
the host version is at least RS5.

Signed-off-by: Nashwan Azhari <nazhari@cloudbasesolutions.com>