oci: appendOSMounts(): remove unused error, and move
This function was added in ae22854e2bfdf233567831d99685a35ef1e4c1af, but never
returned an error, and the error-return was not handled on the callsite. This
patch removes the unused error return, and moves it to a file related to mounts,
which allowed for some of the stubs to be removed and shared between non-FreeBSD
platforms.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>